requirements.txt
absl-py==0.7.1
addict==2.2.1
aiodns==1.1.1
aiohttp==2.3.10
alabaster==0.7.8
albumentations==0.4.1
altgraph==0.15
anaconda-client==1.4.0
anaconda-navigator==1.2.1
argcomplete==1.0.0
astor==0.7.1
astroid==2.0.4
astropy==1.2.1
async-timeout==2.0.1
Babel==2.3.3
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
basemap==1.2.0
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.0
boto==2.40.0
Bottleneck==1.1.0
brewer2mpl==1.4.1
builtwith==1.3.2
Cartopy==0.16.0
cchardet==2.1.4
certifi==2017.4.17
cffi==1.13.0
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.3.7
comtypes==1.1.2
conda==4.5.11
conda-build==1.21.3
configobj==5.0.6
contextlib2==0.5.3
cryptography==2.8
cssselect==1.0.1
cycler==0.10.0
Cython==0.24
cytoolz==0.8.0
dask==0.10.0
datashape==0.5.2
decorator==4.4.0
dill==0.2.5
dm-sonnet==1.33
docutils==0.12
dynd===c328ab7
easydict==1.9
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.11.1
Flask-Cors==2.1.2
future==0.18.2
gast==0.2.2
gdal==2.2.4
geos==0.2.1
gevent==1.1.1
google-pasta==0.2.0
graph-nets==1.0.4
graphviz==0.8.1
greenlet==0.4.10
grpcio==1.19.0
gym==0.15.4
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.9999999
idna==2.5
idna-ssl==1.1.0
imagecorruptions==1.0.0
imageio==2.1.2
imagesize==0.7.1
imgaug==0.2.6
imutils==0.4.5
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
libtiff==0.4.2
llvmlite==0.11.0
locket==0.2.0
lxml==3.8.0
macholib==1.9
Markdown==3.1
MarkupSafe==0.23
matplotlib==2.1.1
mccabe==0.6.1
menuinst==1.4.1
mistune==0.7.2
mmcv==0.2.14
# Editable install with no version control (mmdet==1.0rc0+unknown)
-e g:\expriment\mmdetection
mock==2.0.0
moviepy==0.2.3.2
mpmath==0.19
multidict==4.4.2
multipledispatch==0.4.8
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==2.4
nltk==3.2.1
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.6
numpy==1.17.4
odo==0.5.0
opencv-python==4.1.2.30
openpyxl==2.3.2
pandas==0.18.1
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pbr==5.1.3
pefile==2017.11.5
pep8==1.7.0
pickleshare==0.7.2
Pillow==5.0.0
ply==3.8
prettyplotlib==0.1.7
protobuf==3.10.0
psutil==4.3.0
py==1.4.31
py2exe==0.9.2.2
pyasn1==0.1.9
pycares==2.3.0
pyclipper==1.1.0.post1
pycocotools==2.0
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pydotplus==2.0.2
pyflakes==1.2.3
pyglet==1.3.2
Pygments==2.1.3
PyInstaller==3.3.1
pylint==2.1.1
pyOpenSSL==16.2.0
pyparsing==2.2.0
pyproj==1.9.6
PyQt5==5.9.2
pyquery==1.4.0
pyreadline==2.1
pyshp==2.0.1
pytest==2.9.2
python-dateutil==2.6.1
python-whois==0.6.5
pytz==2017.3
PyWavelets==1.1.1
pywin32==220
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.1
QtPy==1.0.2
requests==2.18.1
rope-py3k==0.9.4.post1
Rtree==0.8.3
ruamel-yaml==0.15.71
scikit-fuzzy==0.3.1
scikit-image==0.15.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
selenium==3.14.1
semantic-version==2.6.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.6
six==1.13.0
sklearn==0.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
splinter==0.9.0
spyder==2.3.9
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
tensorboard==1.9.0
tensorboardX==1.9
tensorflow-estimator==1.14.0
tensorflow-gpu==1.9.0
tensorflow-probability==0.6.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminaltables==3.1.0
tflearn==0.3.2
tifffile==0.14.0
toolz==0.8.0
torchvision==0.4.1+cu92
tornado==4.3
tqdm==4.11.2
traitlets==4.2.1
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.21.1
vboxapi==1.0
webcolors==1.7
Werkzeug==0.12.2
wrapt==1.12.1
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
yarl==1.2.0